audio-wake/requirements.txt
sounddevice
numpy
RPi.GPIO
evdev
